+++ src/app/ui/footer/index.tsx
@@ -4,8 +4,8 @@ import { useState } from 'react';
 import { DEVELOPERS } from './lib/constants';
 import { FOOTER_TEXTS } from './locales';
 
+import { cn } from '@/shared';
 import { RsSchool } from '@/shared/assets/icons';
-import { cn } from '@/shared/lib/utils';
 import { HEADERS_TEXT } from '@/shared/model/constants';
 
 export const Footer = () => {
