+++ src/pages/landing/hooks/index.ts
@@ -1,0 +1,1 @@
+export * from './use-intersection-observer';
