+++ src/shared/index.ts
@@ -1,4 +1,6 @@
-export * from './config/routes';
+export { cn } from './lib/utils';
+
+export { ROUTES } from './config/routes';
 
 export { useTheme } from './hooks/use-theme';
