+++ src/shared/model/constants.ts
@@ -1,9 +1,3 @@
-export const BUTTON_TEXT = {
-  LOGIN: 'Войти',
-  START_FREE: 'Начать бесплатно',
-  REGISTER_NOW: 'Зарегистрироваться сейчас',
-};
-
 export const HEADERS_TEXT = {
   SITE_TITLE: 'Тренажёр интервью',
 };
