+++ src/shared/ui/progress.tsx
@@ -1,7 +1,7 @@
 import { Progress as ProgressPrimitive } from 'radix-ui';
 import * as React from 'react';
 
-import { cn } from '@/shared/lib/utils';
+import { cn } from '@/shared';
 
 function Progress({
   className,
